Hide dependency injection in separate class
All of the APIs use to have a datagen function that can be overridden.
I didn't like that, since it showed up in the documentation.  I opted to
have a class that has all of the data generation functions.  This class
can be injected with an undocumented API in each class.