API to get available free agents (#3)
This adds the ability to get free agent information from your Yahoo!
league.  It is added to the league class.  You specify the position (as
a short code like '2B', 'SS') or position type (e.g. 'B' for all batters
or 'P' for all pitchers).

A list of all free agents for the position is returned.  For each player
we return:
  - name
  - Yahoo! player ID
  - list of eligible positions
  - position type ('B' for batter or 'P' for pitchers)