Fix week_date_range when used during playoffs
Prior to this, if you tried to use week_date_range for a future week in
the playoffs you would get an error.  This is because Yahoo! API only
provide the week date range if the matchups for the week are known.  In
the playoffs this isn't possible.  This commit will now restrict any
calls to week_date_range to be at most one week past the current week.