feat(utils.py): add create_element function #trade_proposal
A new function called 'create_element' has been added to utils.py. This function is used to create an XML element with a given name and text, and append it to a parent element, which will be used to reduce code redundancy when implementing _construct_trade_proposal_xml. This can be used to refactor other xml construction methods as well, but that is outside the scope of this PR.