Include Total Points in player_stats return (#52)
I have updated the player_stats command to leverage the League resource
from the Yahoo API instead of the Player resource. This allows us to
return the total points and all the stats details. There are a few other
benefits IMO:

- Makes the library more predictable as the league class is now using
the league resource instead of an edge case where it uses the player
resource
- We are only gaining data here the league /stats endpoint returns the
same values as the players /stats endpoint
- Allows us to simplify the mock for player_stats

---------

Co-authored-by: DMcP89 <DMcP89@users.noreply.github.com>